3.2
Inside the main function, I created an instance of the struct with your personal birth date details. Modify the infinite loop to cause the LCD on your micro to increment until your age is displayed reach your age. I also corrected 3.1. and which created a temporary instance of the age_data struct with my personal details.